fix compilation error
